Fill in more of implementation

+++ HtmlFormsServer/HtmlFormsServer.h
@@ -15,7 +15,7 @@ FOUNDATION_EXPORT const unsigned char HtmlFormsServerVersionString[];
 
 @protocol HtmlFormsServerDelegate <NSObject>
 
--(void)openUrl:(NSURL*_Nonnull)url windowId:(int)windowId;
+-(void)openUrl:(NSURL*_Nonnull)url windowId:(NSInteger)windowId;
 
 @end
